requirements.txt
twilio~=9.4.2
requests==2.32.3
python-dotenv~=1.0.1
beautifulsoup4~=4.12.3
firebase-admin
pytz~=2024.2
Werkzeug==2.2.2
Flask~=2.1.2
google-cloud-functions
google-cloud-tasks
